errors: better check validity of ErrorCode, ErrorType and RequestErrorType
